% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{read_output}
\alias{read_output}
\title{Read VP results for MCMC analysis}
\usage{
read_output(out_path, out_filename = "vp_results.txt")
}
\arguments{
\item{out_path}{Directory to look for results files in (e.g. C:/VarroaPop/output/)}

\item{out_filename}{Name of the VarroaPop results file to read (e.g. "myresults.txt")}
}
\value{
a dataframe of VP outputs (columns) by simulation date (rows)
}
\description{
Function to read results from a set of 10 vp runs (one for each site/neonic profile)
outputs a 3d array of results (dim 1 = day, dim 2 = response var, dim 3 = site)
}
\author{
Jeffrey M Minucci, \email{jminucci2@gmail.com}
}
