% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_w_exe.R
\name{run_vp}
\alias{run_vp}
\title{Run VarroaPop simulations}
\usage{
run_vp(exe_file, vrp_file, in_file, out_path, out_filename = "vp_results.txt",
  log_path = NULL, logs = T, verbose = F)
}
\arguments{
\item{exe_file}{Full path to the VarroaPop.exe file (e.g. 'C:/VarroaPop/bin/VarroaPop.exe').}

\item{vrp_file}{Full path to the .vrp file (e.g. 'C:/VarroaPop/bin/default.vrp').}

\item{in_file}{Full path to the input file (e.g. 'C:/VarroaPop/input/myfile.txt').}

\item{out_path}{Full path to the output folder, with trailing slash (e.g. 'C:/VarroaPop/output/).}

\item{out_filename}{Optional: custom file name for the output file.}

\item{log_path}{path to the log folder (e.g. 'C:/VarroaPop/logs).(can be omitted if logs=F)}

\item{logs}{T/F, should VP write a log file to the log_path folder?}

\item{verbose}{T/F, print extra details?}
}
\value{
None - VP writes output to the out_path directory
}
\description{
Wrapper code to run the VP exe for MCMC analysis.
}
\author{
Jeffrey M Minucci, \email{jminucci2@gmail.com}
}
