% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_varroapop.R
\name{RunVarroaPop}
\alias{RunVarroaPop}
\title{Run the VarroaPop honey bee colony model}
\usage{
RunVarroaPop(parameters)
}
\arguments{
\item{parameters}{Named vector or list of VarroaPop inputs to be written to .txt file.}
}
\value{
A dataframe where each column is a VarroaPop output and rows are daily values (with the first row
 giving initial values at time 0.)
}
\description{
Function to run the VarroaPop honey bee colony model for a given set of parameters. Currently using
default weather from Columbus, Ohio 1990-2015.
}
\examples{
 parameters <- c("ICWorkerBrood"=1000, "ICWorkerAdult"=1500)
 RunVarroaPop(parameters)
}
\author{
Jeffrey M Minucci, \email{jminucci2@gmail.com}
}
