% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_varroapop.R
\name{RunVarroaPopLocal}
\alias{RunVarroaPopLocal}
\title{Run the VarroaPop honey bee colony model on a local computer}
\usage{
RunVarroaPopLocal(parameters, exe_file = system.file("varroapop_files", "exe",
  "VarroaPop.exe", package = "VarroaPopWrapper"),
  vrp_file = system.file("varroapop_files", "exe", "default.vrp", package =
  "VarroaPopWrapper"), in_path = system.file("varroapop_files", "input",
  package = "VarroaPopWrapper"), in_filename = "vp_input.txt",
  log_path = system.file("varroapop_files", "logs", package =
  "VarroaPopWrapper"), out_path = system.file("varroapop_files", "output",
  package = "VarroaPopWrapper"), out_filename = "vp_results.txt",
  save_files = FALSE, logs = FALSE, verbose = FALSE)
}
\arguments{
\item{parameters}{Named vector of VarroaPop inputs to be written to .txt file.}

\item{exe_file}{= Full path to the VarroaPop exe file, e.g. 'C:/VarroaPop/exe/VarroaPop.exe' (optional)}

\item{vrp_file}{= Full path to the VarroaPop session file, e.g. 'C:/VarroaPop/exe/default.vrp' (optional)}

\item{in_path}{Full path of directory to write inputs file to, e.g. 'C:/VarroaPop/inputs/' (optional)}

\item{in_filename}{Name of input files which will be created e.g. 'myinput.txt' (optional)}

\item{log_path}{Full path of directory to write log files to e.g. 'C:/VarroaPop/logs/'(optional)}

\item{out_path}{Full path of directory to write outputs file to, e.g. 'C:/VarroaPop/outputs/' (optional)}

\item{out_filename}{Name of input files which will be created e.g. 'myoutput.txt' (optional)}

\item{save_files}{T/F should input, output, and log files be kept or deleted afterward?}

\item{logs}{T/F create VarroaPop program log files?}

\item{verbose}{T/F print extra details?}
}
\value{
A dataframe where each column is a VarroaPop output and rows are daily values (with the first row
 giving initial values at time 0.)
}
\description{
Function to run the VarroaPop honey bee colony model for a given set of parameters, with options to set and
 save the input, output, and log files. Path to the weather file
 must be set in the .vrp file, or as the input parameter 'WeatherFileName'
}
\author{
Jeffrey M Minucci, \email{jminucci2@gmail.com}
}
