% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{write_vp_input}
\alias{write_vp_input}
\title{Write a VarroaPop input file from a named list or vector}
\usage{
write_vp_input(params, in_path = system.file("varroapop_files", "input",
  package = "VarroaPopWrapper"), in_filename = "vp_input.txt", verbose = F)
}
\arguments{
\item{params}{Named vector of VarroaPop inputs to be written to .txt file.}

\item{in_path}{Directory to write vp_input.txt file to (optional).}

\item{in_filename}{Filename of the written input file. Defaults to 'vp_input.txt'.}

\item{verbose}{T/F print extra details?}
}
\value{
None... writes inputs to a .txt file in in_path for VarroaPop
}
\description{
Function to create a single input file from a set of parameters in the form of
a one row dataframe, where columns are named.
}
\examples{
 parameters <- c("foo"=15,"bar"=4)
 write_vp_input(parameters, in_path = "d:/path/to/inputdir")

}
\author{
Jeffrey M Minucci, \email{jminucci2@gmail.com}
}
