% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_input.R
\name{write_vp_input_long}
\alias{write_vp_input_long}
\title{Write a VarroaPop input file where each named parameter is to be written (except in_path or verbose)}
\usage{
write_vp_input_long(in_path = paste(system.file(package = "VarroaPopWrapper"),
  "/varroapop_files/input", sep = ""), ..., in_filename = "vp_input.txt",
  verbose = F)
}
\arguments{
\item{in_path}{Directory to write vp_input.txt file to}

\item{...}{Initial parameters to be passed to VarroaPop}

\item{in_filename}{Filename of the written input file. Defaults to 'vp_input.txt'.}

\item{verbose}{T/F print extra details?}
}
\value{
None... writes inputs to a .txt file in in_path for VarroaPop
}
\description{
function to create a single input file from a set of parameters in the form of
a one row dataframe, where columns are named
}
\examples{
 parameters <- c()
 write_vp_input(in_path = "d:/path/to/inputdir",foo=15,bar=4)

}
